reducers/ResourceValidationHelpers refactored out of ResourceHelpers
